+++ applications/track-manager/pom.xml
@@ -2,7 +2,7 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.hashmapinc.tempus.edge</groupId>
   <artifactId>track-manager</artifactId>
-  <version>0.1.0</version>
+  <version>0.1.1</version>
   <name>${project.artifactId}</name>
   <description>Tempus Edge application for managing track ID, external configs, and ioFog configs for a whole track.</description>
   <inceptionYear>2018</inceptionYear>
@@ -32,13 +32,13 @@
       <version>1.7.25</version>
     </dependency>
     <dependency>
-      <groupId>com.iotracks</groupId>
-      <artifactId>container-java-sdk</artifactId>
-      <version>1.1.3</version>
+      <groupId>com.hashmapinc.tempus.edge</groupId>
+      <artifactId>protos</artifactId>
+      <version>0.1.0</version>
     </dependency>
     <dependency>
       <groupId>com.hashmapinc.tempus.edge</groupId>
-      <artifactId>protos</artifactId>
+      <artifactId>iofog</artifactId>
       <version>0.1.0</version>
     </dependency>
     <dependency>
